complete waiter api and some fixes
* optimise create order (get only last 24 hours orders)
* remove commented test
* add some exceptions with exception handling
* code clean up

+++ src/main/java/com/sflpro/cma/repository/TableRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Set;
 
-public interface TableRepository extends JpaRepository<Table, Long> {
+public interface TableRepository extends JpaRepository<Table, Integer> {
 
     Set<Table> findByUserId(Long userId);
